% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addunderscore.R
\name{addunderscore}
\alias{addunderscore}
\title{Add Underscore}
\usage{
addunderscore(term)
}
\arguments{
\item{term}{a character string with punctuation and spaces.}
}
\value{
a character
}
\description{
A function that removes punctuation and spaces from a character string
}
\examples{
addunderscore("St. James")

}
\seealso{
Other Sentinel Lakes Tools: 
\code{\link{fixlakeid}()},
\code{\link{lakeid2name}()},
\code{\link{lakename2id}()}
}
\concept{Sentinel Lakes Tools}
