% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_do_folder.R
\name{compile_do_folder}
\alias{compile_do_folder}
\title{Compile DO data}
\usage{
compile_do_folder(folderpath, savetofolder)
}
\arguments{
\item{folderpath}{The folder where the individual files are located}

\item{savetofolder}{The folder where the output csv is to be saved}
}
\value{
a .csv file
}
\description{
This function compiles Dissolved Oxygen data files downloaded from PME miniDot DO loggers. Use if the data where not compiled using the PME java app on the logger.
}
\examples{
\dontrun{
compile_do_folder(
  folderpath = "C:/Projects/Datasets/DO/Continuous/Greenwood/2019/7450-516388",
  savetopath = "C:/Projects/Datasets/DO/Continuous/Greenwood/2019")
}
}
\keyword{DO}
\keyword{Dissolved}
\keyword{Oxygen}
