% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuousdatameans.R
\name{continuousdatameans}
\alias{continuousdatameans}
\title{Continous Lake Data Summary}
\usage{
continuousdatameans(contdata, timedata, stratifieddata,
  period = c("daily" | "weekly" | "monthly" | "yearly"))
}
\arguments{
\item{contdata}{vector with continuous numeric data. Must have the same number of values as the timedata and stratifieddata parameters.}

\item{timedata}{vector with date and/or time data in POSIXcT format. Must have the same number of values as the contdata and stratifieddata values.}

\item{stratifieddata}{vector with additional data that indicates how the contdata should be stratified. For example, depth data for water temperature. Must have the same number of values as the contdata and timedata values.}

\item{period}{the time period to calculate the mean. Choices include daily, weekly, monthly, or yearly.}
}
\value{
a data.frame with averaged continuous data.
}
\description{
This function calculates the mean value of a continuous data set for a specified time period.
}
\examples{
\dontrun{

x <-continuousdatasummary(
     contdata = temperaturedata,
     timedata = date_time,
     stratifieddata = logger_depths,
     period = "daily")

}
}
\keyword{lake}
\keyword{profile}
\keyword{temperature}
\keyword{thermal}
