% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2sentinel.R
\name{fish2sentinel}
\alias{fish2sentinel}
\title{Fish Survey Data Sentinel Formatting}
\usage{
fish2sentinel(fishsurvey)
}
\arguments{
\item{fishsurvey}{results from the fishsurveydata() function.}
}
\description{
Reorganizes fish survey data downloaded from lakefinder into a common Sentinel Lakes format, making it easier to integrate with other Sentinel Lakes 
formatted datasets.
}
\examples{
#' #Retrieve the lakefinder data
x <- lakefinderdownload("11041300")

#Extract the fish survey data
y <- fishsurveydata(x)

#Convert to Sentinel Lakes formatting
z <- fish2sentinel(y)

}
\seealso{
Other Formatting: \code{\link{wq2sentinel}}
}
\concept{Formatting}
\keyword{fish}
\keyword{format}
\keyword{sentinel}
