% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishabbrev2common.R
\name{fishabbrev2common}
\alias{fishabbrev2common}
\title{Fish Abbreviations to Common Names}
\usage{
fishabbrev2common(fishabbreviation)
}
\arguments{
\item{fishabbreviation}{a three character string}
}
\value{
a character
}
\description{
A function that returns the common name of a fish based upon the fish abbreviation provided.
}
\examples{
fishabbrev2common("BLG")

}
\seealso{
Other Fish: 
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
