% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishabbreviations.R
\docType{data}
\name{fishabbreviations}
\alias{fishabbreviations}
\title{Fish Species Abbreviations}
\format{a data frame.}
\usage{
data(fishabbreviations)
}
\description{
A lookup table connecting fish species common names with MN DNR species name abbreviations.
}
\keyword{datasets}
