% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishcommon2abbrev.R
\name{fishcommon2abbrev}
\alias{fishcommon2abbrev}
\title{Fish Common Names to Abbreviations}
\usage{
fishcommon2abbrev(commonname)
}
\arguments{
\item{commonname}{a character string}
}
\value{
a character
}
\description{
A function that returns the abbreviation of a fish based upon the common name provided.
}
\examples{
fishcommon2abbrev("Bluegill")

}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
