% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishcommon2scientific.R
\name{fishcommon2scientific}
\alias{fishcommon2scientific}
\title{Fish Common Names to Scientific Names}
\usage{
fishcommon2scientific(commonname)
}
\arguments{
\item{commonname}{a character string}
}
\value{
a character
}
\description{
A function that returns the scientific name of a fish based upon the common name provided.
}
\examples{
fishcommon2scientific("Bluegill")

}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
