% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishscientific2abbrev.R
\name{fishscientific2abbrev}
\alias{fishscientific2abbrev}
\title{Fish Scientific Names to Abbreviations}
\usage{
fishscientific2abbrev(scientificname)
}
\arguments{
\item{scientificname}{a character string}
}
\value{
a character
}
\description{
A function that returns the abbreviation of a fish based upon the scientific name provided.
}
\examples{
fishscientific2abbrev("Lepomis macrochirus")

}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
