% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishscientific2common.R
\name{fishscientific2common}
\alias{fishscientific2common}
\title{Fish Scientific Names to Common Names}
\usage{
fishscientific2common(scientificname)
}
\arguments{
\item{scientificname}{a character string}
}
\value{
a character
}
\description{
A function that returns the common name(s) of a fish based upon the scientific name provided.
}
\examples{
fishscientific2common("Lepomis macrochirus")

}
\seealso{
Other Fish: \code{\link{fishabbrev2common}},
  \code{\link{fishabbrev2scientific}},
  \code{\link{fishcommon2abbrev}},
  \code{\link{fishcommon2scientific}},
  \code{\link{fishscientific2abbrev}},
  \code{\link{fishspeciesmetadata}},
  \code{\link{fishsurveydata}}, \code{\link{fishtable}},
  \code{\link{fishtrendplots}},
  \code{\link{fishtrendstats}},
  \code{\link{lakefinderdownload}}
}
\concept{Fish}
