% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishspeciesmetadata.R
\docType{data}
\name{fishspeciesmetadata}
\alias{fishspeciesmetadata}
\title{Fish Species Metadata}
\format{a data frame.}
\usage{
data(fishspeciesmetadata)
}
\description{
A dataset that includes name abbreviations and most effective gear for a selection of fish species.
}
\seealso{
Other Fish: \code{\link{fishsurveydata}},
  \code{\link{fishtable}}, \code{\link{fishtrendplots}},
  \code{\link{fishtrendstats}},
  \code{\link{lakefinderdownload}}
}
\concept{Fish}
\keyword{datasets}
