% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishsurveydata.R
\name{fishsurveydata}
\alias{fishsurveydata}
\title{Lakefinder Fish Survey Data}
\usage{
fishsurveydata(lakefinderdata)
}
\arguments{
\item{lakefinderdata}{results from the readlakefinder() function.}
}
\value{
a data.frame of fish survey data
}
\description{
A function to extract fish survey data from the data downloaded using the readlakefinder() function.
}
\examples{
#Retrieve the lakefinder data
x <- lakefinderdownload("11041300")
#Extract the fish survey data
y <- fishsurveydata(x)

}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
\keyword{data}
\keyword{fish}
\keyword{lakefinder}
