% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishtable.R
\name{fishtable}
\alias{fishtable}
\title{Fish Data table}
\usage{
fishtable(fishsurvey, fishspecies, startyear = NULL, endyear = NULL)
}
\arguments{
\item{fishsurvey}{results from the fishsurveydata() function.}

\item{fishspecies}{a list of fish species names present in the fishspeciesmetadata table. Valid species include: "White Sucker", "Black Crappie", "Bluegill",
"Largemouth Bass", "Rock Bass", "Smallmouth Bass", "Muskellunge", "Northern Pike", "Black Bullhead", "Brown Bullhead", "Channel Catfish", "Yellow Bullhead",
"White Bass", "Walleye", "Yellow Perch", "Cisco", and "Lake Trout".}

\item{startyear}{a numeric of the earliest year under consideration in the YYYY format. Default is NULL.}

\item{endyear}{a numeric of the latest year under consideration in the YYYY format. Default is NULL.}
}
\value{
a data.frame of fish catch data
}
\description{
This function selects fish catch data for the appropriate gear for the list of fish species in the fishmetadata table.
}
\details{
Some surveys use Shallow and Deep gillnets as opposed to Standard gillnets. When this occurs, the total catch between those two gear types will be added and divided by the
gear count.

Some data have different units for total weights and average weights. In some cases both measurements are in pounds, but in other cases total weights are in grams and
average weights are in pounds. Currently this function maintains these differences.
}
\examples{
\dontrun{
#Retrieve the lakefinder data
x <- lakefinderdownload("11041300")
#Extract the fish survey data
y <- fishsurveydata(x)
#Select the fish catch data
z <-fishtable(
      fishsurvey = y,
      fishspecies = c("Bluegill","Largemouth Bass"))
}
}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
\keyword{fish}
\keyword{gear}
\keyword{lakefinder}
\keyword{survey}
