% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishtrendplots.R
\name{fishtrendplots}
\alias{fishtrendplots}
\title{Fish Trend Plots}
\usage{
fishtrendplots(plotdata, logtransform = TRUE, saveto = paste0(getwd()),
  ptcolor = "blue", ptsize = 3, pttype = 20, addlm = TRUE,
  lmcolor = "black", lmsize = 1, lmtype = 2, includestats = TRUE,
  statstable = TRUE, maxpvalue = 0.1)
}
\arguments{
\item{plotdata}{a water quality data.frame created with the fishtable() function.}

\item{logtransform}{a logical indicating whether the statistics should be calulated using the natural log of the CPUE +1 values. Defaults to TRUE.}

\item{saveto}{designates the folder where the plots and table should be saved. Defaults to the working directory.}

\item{ptcolor}{designates the color of the points in the plots. Defaults to "blue".}

\item{ptsize}{designates the size of the points in the plots. Defaults to 3.}

\item{pttype}{designates the type of the points in the plots. Defaults to 20.}

\item{addlm}{a logical indicating whether the linear regression line should be displayed on the plot. Defaults to TRUE.}

\item{lmcolor}{designates the color of the linear regression line. Defaults to "black".}

\item{lmsize}{designates the size of the linear regression line. Defaults to 1.}

\item{lmtype}{designates the type of the linear regression line. Defaults to 2.}

\item{includestats}{a logical indicating whether the R-squared and p-value statistics text should be displayed on the plot. Defaults to TRUE.}

\item{statstable}{a logical indicating whether a .csv of trend statistics should be saved. Defaults to TRUE.}

\item{maxpvalue}{a numeric indicating the maximum p-value limit for a species' CPUE trend to be included in the export. Any species' CPUE trends with a
p-value above this number will not be exported. Defaults to 0.1.}
}
\value{
.png plots and statistics .csv
}
\description{
This function creates and saves fish trend scatter plots comparing CPUE across surveys. It will also save a .csv table with trend statistics calculated with
the fishstats() function.
}
\examples{
#Download the data
x <- lakefinderdownload("21005700")

#Select fisheries survey data
y <- fishsurveydata(
       lakefinderdata = x
     )

#Extract appropriate data for each fish species
z <-fishtable(
      fishsurvey = y,
      fishspecies = c("Bluegill","Largemouth Bass"))

fishtrendplots(plotdata = z)

}
\seealso{
Other Fish: \code{\link{fishabbrev2common}},
  \code{\link{fishabbrev2scientific}},
  \code{\link{fishcommon2abbrev}},
  \code{\link{fishcommon2scientific}},
  \code{\link{fishscientific2abbrev}},
  \code{\link{fishscientific2common}},
  \code{\link{fishspeciesmetadata}},
  \code{\link{fishsurveydata}}, \code{\link{fishtable}},
  \code{\link{fishtrendstats}},
  \code{\link{lakefinderdownload}}
}
\concept{Fish}
\keyword{fish}
\keyword{trends}
