% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishtrendstats.R
\name{fishtrendstats}
\alias{fishtrendstats}
\title{Calculate Fish CPUE Linear Trend Statistics}
\usage{
fishtrendstats(statdata, logtransform = TRUE)
}
\arguments{
\item{statdata}{a fish species data.frame processed through the fishtable() function.}

\item{logtransform}{a logical indicating whether the statistics should be calculated using the natural log of the CPUE + 1 values.  Defaults to TRUE.}
}
\value{
a data.frame with statistics for each species' CPUE.
}
\description{
This function calculates linear trend statistics using data download from the MNDNR Lakefinder website using the output of the fishtable() function.
}
\examples{
\dontrun{
#Download the data
x <- lakefinderdownload("21005700")

#Select fisheries survey data
y <- fishsurveydata(
       lakefinderdata = x
     )

#Extract appropriate data for each fish species
z <-fishtable(
      fishsurvey = y,
      fishspecies = c("Bluegill","Largemouth Bass"))

fishtrendstats(statdata = z)
}
}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{lakefinderdownload}()}
}
\concept{Fish}
\keyword{analysis}
\keyword{fish}
\keyword{minnesota}
\keyword{mndnr}
\keyword{natural}
\keyword{resources}
\keyword{statistics}
\keyword{trend}
