% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixlakeid.R
\name{fixlakeid}
\alias{fixlakeid}
\title{Fix Minnesota Lake IDs}
\usage{
fixlakeid(lakeid)
}
\arguments{
\item{lakeid}{lakeid number}
}
\value{
a character
}
\description{
This function ensures Minnesota Lake IDs (also known as DOWLKNUMs) are characters with "0" at the front and do not have any dashes.
}
\examples{
fixlakeid(6000200)
fixlakeid("06-0002-00")

}
\seealso{
Other Sentinel Lakes Tools: \code{\link{addunderscore}},
  \code{\link{gpx2shp}}, \code{\link{lakeidtoname}},
  \code{\link{lakenametoid}},
  \code{\link{readaccessdatabase}}
}
\concept{Sentinel Lakes Tools}
\keyword{dowlknum}
\keyword{lakeid}
