% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icedownload.R
\name{icedownload}
\alias{icedownload}
\title{Download Minnesota Climatology Lake Ice Data}
\usage{
icedownload(lakeid)
}
\arguments{
\item{lakeid}{a character indicating the LakeId (DOWLKNUM) for the lake to be downloaded.}
}
\value{
a data.frame with lake ice data.
}
\description{
This function downloads lake ice in and out dates from the Minnesota Climatology website: https://www.dnr.state.mn.us/ice_out/index.html and
https://www.dnr.state.mn.us/ice_in/index.html. The data are already formatted into the Sentinel Lakes format.
}
\concept{Ice}
\keyword{climate}
\keyword{data}
\keyword{download}
\keyword{ice}
