% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakefinderdownload.R
\name{lakefinderdownload}
\alias{lakefinderdownload}
\title{Download Lakefinder Data}
\usage{
lakefinderdownload(lakeid)
}
\arguments{
\item{lakeid}{a character indicating the LakeId (DOWLKNUM) for the lake to be downloaded.}
}
\value{
a list of data
}
\description{
This function downloads the MNDNR Lakefinder website json tables for the selected lake. It was developed by user hrbrmstr at https://stackoverflow.com/questions/46517463/scraping-html-data-table-using-rvest.
}
\examples{
x <- lakefinderdownload("11041300")

}
\seealso{
Other Fish: 
\code{\link{fishabbrev2common}()},
\code{\link{fishabbrev2scientific}()},
\code{\link{fishcommon2abbrev}()},
\code{\link{fishcommon2scientific}()},
\code{\link{fishscientific2abbrev}()},
\code{\link{fishscientific2common}()},
\code{\link{fishspeciesmetadata}},
\code{\link{fishsurveydata}()},
\code{\link{fishtable}()},
\code{\link{fishtrendplots}()},
\code{\link{fishtrendstats}()}
}
\concept{Fish}
\concept{Lake}
\keyword{data}
\keyword{lakefinder}
\keyword{minnesota}
\keyword{natural}
\keyword{resources}
\keyword{survey}
