% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeid2name.R
\name{lakeid2name}
\alias{lakeid2name}
\title{LakeId to Lake Name}
\usage{
lakeid2name(lakeid)
}
\arguments{
\item{lakeid}{LakeId number.}
}
\value{
a character
}
\description{
This function returns the name of the lake for the given LakeId.
}
\examples{
lakeid2name("02000400")

}
\seealso{
Other Sentinel Lakes Tools: \code{\link{addunderscore}},
  \code{\link{fixlakeid}}, \code{\link{gpx2shp}},
  \code{\link{lakename2id}},
  \code{\link{readaccessdatabase}}
}
\concept{Sentinel Lakes Tools}
\keyword{dowlknum}
\keyword{lakeid}
\keyword{lakename}
