% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeidtoname.R
\name{lakeidtoname}
\alias{lakeidtoname}
\title{LakeId to Lake Name}
\usage{
lakeidtoname(lakeid)
}
\arguments{
\item{lakeid}{LakeId number.}
}
\value{
a character
}
\description{
This function returns the name of the lake for the given LakeId.
}
\examples{
lakeidtoname("02000400")

}
\seealso{
Other Sentinel Lakes Tools: \code{\link{addunderscore}},
  \code{\link{fixlakeid}}, \code{\link{gpx2shp}},
  \code{\link{lakenametoid}},
  \code{\link{readaccessdatabase}}
}
\concept{Sentinel Lakes Tools}
\keyword{dowlknum}
\keyword{lakeid}
\keyword{lakename}
