% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakename2id.R
\name{lakename2id}
\alias{lakename2id}
\title{Lake Name to LakeIds}
\usage{
lakename2id(lakename, county = NULL)
}
\arguments{
\item{lakename}{name of the lake.}

\item{county}{county where the lake is located. Default is NULL.}
}
\value{
a data.frame
}
\description{
This function creates a data frame with the LakeIds for all lakes with the input name, as well as the County they are located in. If multiple lakes have the
same name, this will return a list of LakeIds.
}
\examples{
lakename2id("Echo")

}
\seealso{
Other Sentinel Lakes Tools: \code{\link{addunderscore}},
  \code{\link{fixlakeid}}, \code{\link{gpx2shp}},
  \code{\link{lakeid2name}}
}
\concept{Sentinel Lakes Tools}
\keyword{dowlknum}
\keyword{lakeid}
