% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlakesmetadata.R
\docType{data}
\name{mnlakesmetadata}
\alias{mnlakesmetadata}
\title{Minnesota Lakes Metadata}
\format{
a data frame.
}
\usage{
data(mnlakesmetadata)
}
\description{
A dataset with metadata for Minnesota lakes with a lakeid (DOWLKNUM).
}
\keyword{datasets}
