% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnpcastations.R
\docType{data}
\name{mnpcastations}
\alias{mnpcastations}
\title{MNPCA Stations}
\format{a data frame.}
\usage{
data(mnpcastations)
}
\description{
A dataset with lists of mnpca stations for each Sentinel Lake.
}
\keyword{datasets}
