% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readaccessdatabase.R
\name{readaccessdatabase}
\alias{readaccessdatabase}
\title{Read Access Databases}
\usage{
readaccessdatabase(database, sqtable)
}
\arguments{
\item{database}{file path to the Access database.}

\item{sqtable}{name of the table in the database.}
}
\value{
data.frame
}
\description{
This function opens a connection to an Access database, reads the selected table, and closes the connection using the RODBC R package.
}
\seealso{
Other Sentinel Lakes Tools: \code{\link{addunderscore}},
  \code{\link{fixlakeid}}, \code{\link{gpx2shp}},
  \code{\link{lakeidtoname}}, \code{\link{lakenametoid}}
}
\concept{Sentinel Lakes Tools}
\keyword{access}
\keyword{database}
\keyword{import}
\keyword{read}
