% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinellakesmetadata.R
\docType{data}
\name{sentinellakesmetadata}
\alias{sentinellakesmetadata}
\title{Sentinel Lakes Metadata}
\format{a data frame.}
\usage{
data(sentinellakesmetadata)
}
\description{
A dataset that contains various metadata required for the mnsentinellakes package.
}
\keyword{datasets}
