% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinelmnpcastations.R
\docType{data}
\name{sentinelmnpcastations}
\alias{sentinelmnpcastations}
\title{Sentinel MNPCA Stations}
\format{a data frame.}
\usage{
data(sentinelmnpcastations)
}
\description{
A dataset with a list of Sentinel Lake mnpca stations.
}
\keyword{datasets}
