% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather2sentinel.R
\name{weather2sentinel}
\alias{weather2sentinel}
\title{Weather Data Sentinel Formatting}
\usage{
weather2sentinel(weatherdata)
}
\arguments{
\item{weatherdata}{a weather data.frame downloaded using the weatherdownload() function.}
}
\description{
Reorganizes weather data downloaded from Iowa State's Mesonet website into a common Sentinel Lakes format, making it easier to integrate with other Sentinel
Lakes formatted datasets. This will also convert all units to metric. All numbers rounded to 2 decimal places.
}
\examples{
#Download the data
x <- weatherdownload(
       lakeid = 21005700,
       startdate = "2019-03-20",
       enddate = "2019-04-01")
#Convert to Sentinel Lakes formatting
y <- weather2sentinel(x)
}
\seealso{
Other Formatting: 
\code{\link{fish2sentinel}()},
\code{\link{wq2sentinel}()}
}
\concept{Formatting}
\keyword{weather}
