% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherdownload.R
\name{weatherdownload}
\alias{weatherdownload}
\title{Downloads Airport Weather Data}
\usage{
weatherdownload(lakeid, startdate, enddate,
  parameters = c("Air Temperature", "Dew Point", "Relative Humidity",
  "Wind Direction", "Wind Speed", "Altimeter", "Precipitation",
  "Gust Speed"))
}
\arguments{
\item{lakeid}{Minnesota lake identifier (DOWLKNUM) for the lake of interest.}

\item{startdate}{a date indicating the beginning of the date range to be downloaded. Format: "yyyy-mm-dd".}

\item{enddate}{a date indicating the end of the date range to be downloaded. Format: "yyyy-mm-dd".}

\item{parameters}{a list of parameters to download. Valid parameters include "Air Temperature", "Dew Point", "Relative Humidity", "Wind Direction",
"Wind Speed", "Altimeter", "Precipitation", and "Gust Speed". Default includes all parameters.}
}
\description{
This function downloads airport weather station data from the Iowa State Mesonet from the station nearest to chosen lake.
}
\examples{
x <- weatherdownload(
       lakeid = 21005700,
       startdate = "2019-03-20",
       enddate = "2019-04-01")
}
\concept{Weather}
\keyword{air}
\keyword{altimeter}
\keyword{dew}
\keyword{direction}
\keyword{gust}
\keyword{humidity}
\keyword{iowa}
\keyword{mesonet}
\keyword{point}
\keyword{precipitation}
\keyword{relative}
\keyword{speed}
\keyword{state}
\keyword{temperature}
\keyword{weather}
\keyword{wind}
