% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherevents.R
\name{weatherevents}
\alias{weatherevents}
\title{Weather Events}
\usage{
weatherevents(lakeid, startdate, enddate)
}
\arguments{
\item{lakeid}{Minnesota lake identifier (DOWLKNUM) for the lake of interest.}

\item{startdate}{a date indicating the beginning of the date range to be downloaded. Format: "yyyy-mm-dd".}

\item{enddate}{a date indicating the end of the date range to be downloaded. Format: "yyyy-mm-dd".}
}
\description{
Determines which days experienced wind (>5 knots), precipitation, or clouds for more than half the day.
}
\examples{
x = weatherevents(
 lakeid = "16007700",
 startdate = "2019-04-01",
 enddate = "2019-04-30")

}
\keyword{weather}
