% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq2sentinel.R
\name{wq2sentinel}
\alias{wq2sentinel}
\title{Water Quality Data Sentinel Formatting}
\usage{
wq2sentinel(wqdata)
}
\arguments{
\item{wqdata}{a water quality data.frame downloaded using the wqdatadownload() function.}
}
\description{
Reorganizes water quality data downloaded from the MNPCA's EDA website into a common Sentinel Lakes format, making it easier to integrate with other Sentinel 
Lakes formatted datasets.
}
\examples{
#Download the data
x <- wqdatadownload(c("15-0010-00-100","15-0010-00-101","15-0010-00-102"))

#Convert to Sentinel Lakes formatting
y <- wq2sentinel(x)
}
\seealso{
Other Formatting: \code{\link{fish2sentinel}}
}
\concept{Formatting}
\keyword{format}
\keyword{mnpca}
\keyword{quality}
\keyword{sentinel}
\keyword{water}
