% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqdatadownload.R
\name{wqdatadownload}
\alias{wqdatadownload}
\title{Download MNPCA Water Quality Data}
\usage{
wqdatadownload(stationids)
}
\arguments{
\item{stationids}{a character vector of station ids in the format "##-####-##-###".}
}
\value{
data.frame with the downloaded data.
}
\description{
This function allows you to download data directly from the Minnesota Pollution Control Agency's EDA website.
}
\examples{
wqdatadownload("15-0010-00-100")
wqdatadownload(c("15-0010-00-100","15-0010-00-101","15-0010-00-102"))

}
\seealso{
Other Water Quality: 
\code{\link{wqmonthtable}()},
\code{\link{wqmonthtrendplots}()},
\code{\link{wqmonthtrendstats}()},
\code{\link{wqparameters}()},
\code{\link{wqstations}()}
}
\concept{Water Quality}
\keyword{download.}
\keyword{minnesota}
\keyword{mnpca}
\keyword{pollution}
\keyword{quality}
\keyword{water}
