% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqmonthtrendplots.R
\name{wqmonthtrendplots}
\alias{wqmonthtrendplots}
\title{WQ Trend Plots}
\usage{
wqmonthtrendplots(plotdata, logtransform = FALSE,
  saveto = paste0(getwd()), ptcolor = "blue", ptsize = 3,
  pttype = 20, addlm = TRUE, lmcolor = "black", lmsize = 1,
  lmtype = 2, includestats = TRUE, statstable = TRUE,
  maxpvalue = 0.1)
}
\arguments{
\item{plotdata}{a water quality data.frame created with the wqmonthtable() function.}

\item{logtransform}{a logical indicating whether the statistics should be calulated using the natural log of the parameter values. Defaults to FALSE.}

\item{saveto}{designates the folder where the plots and table should be saved. Defaults to the working directory.}

\item{ptcolor}{designates the color of the points in the plots. Defaults to "blue".}

\item{ptsize}{designates the size of the points in the plots. Defaults to 3.}

\item{pttype}{designates the type of the points in the plots. Defaults to 20.}

\item{addlm}{a logical indicating whether the linear regression line should be displayed on the plot. Defaults to TRUE.}

\item{lmcolor}{designates the color of the linear regression line. Defaults to "black".}

\item{lmsize}{designates the size of the linear regression line. Defaults to 1.}

\item{lmtype}{designates the type of the linear regression line. Defaults to 2.}

\item{includestats}{a logical indicating whether the R-squared and p-value statistics text should be displayed on the plot. Defaults to TRUE.}

\item{statstable}{a logical indicating whether a .csv of trend statistics should be saved. Defaults to TRUE.}

\item{maxpvalue}{a numeric indicating the maximum p-value limit for a parameter trend to be included in the export. Any parameters trends with a p-value above this number will not be exported. Defaults to 0.1.}
}
\value{
.png plots and statistics .csv
}
\description{
This function creates and saves water quality trend scatter plots comparing parameters for specific months across years. It will also save a .csv table with
trend statistics calculated with the wqmonthrendstats() function.
}
\examples{
#Download the data
x <- wqdatadownload(c("15-0010-00-100","15-0010-00-101","15-0010-00-102"))

#Process multiple parameters across multiple months within a given year range
y <- wqmonthtable(
       wqdata = x,
       parameters = c("Depth, Secchi disk depth","Temperature, water","pH"),
       months = c(7,8,9),
       startyear = 2008,
       endyear = 2018
     )

#Create the trend plots
wqmonthtrendplots(
  plotdata = y
  )

}
\seealso{
Other Water Quality: \code{\link{wqdatadownload}},
  \code{\link{wqmonthtable}},
  \code{\link{wqmonthtrendstats}},
  \code{\link{wqparameters}}, \code{\link{wqstations}}
}
\concept{Water Quality}
\keyword{quality}
\keyword{trends}
\keyword{water}
