% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqparameters.R
\name{wqparameters}
\alias{wqparameters}
\title{Water Quality Parameters}
\usage{
wqparameters(wqdata, minsample = 0)
}
\arguments{
\item{wqdata}{a water quality data.frame downloaded using the wqdatadownload() function.}

\item{minsample}{the minimum number of parameter samples required to be included in the list. Defaults to 0.}
}
\value{
A vector of water quality parameters.
}
\description{
This function returns a vector of the water quality parameters present in data downloaded from the Minnesota Pollution Control Agency's EDA website.
}
\examples{
x <- wqdatadownload("15-0010-00-100")
wqparameters(x)

}
\seealso{
Other Water Quality: \code{\link{wqdatadownload}},
  \code{\link{wqmonthtable}},
  \code{\link{wqmonthtrendplots}},
  \code{\link{wqmonthtrendstats}}, \code{\link{wqstations}}
}
\concept{Water Quality}
\keyword{minnesota}
\keyword{mnpca}
\keyword{parameters}
\keyword{pollution}
\keyword{quality}
\keyword{water}
