% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqstations.R
\name{wqstations}
\alias{wqstations}
\title{Water Quality Stations}
\usage{
wqstations(lakeid)
}
\arguments{
\item{lakeid}{a character indicating the LakeId (DOWLKNUM) for the lake to be downloaded.}
}
\value{
a list of water quality station ids
}
\description{
This function returns a list of station ids located on the lake with the supplied LakeId.
}
\seealso{
Other Water Quality: \code{\link{wqdatadownload}},
  \code{\link{wqmonthtable}},
  \code{\link{wqmonthtrendplots}},
  \code{\link{wqmonthtrendstats}},
  \code{\link{wqparameters}}
}
\concept{Water Quality}
\keyword{control}
\keyword{minnesota}
\keyword{mnpca}
\keyword{pollution}
\keyword{quality}
\keyword{stations}
\keyword{water}
