% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeaccessdatabase.R
\name{writeaccessdatabase}
\alias{writeaccessdatabase}
\title{Write to Access Database}
\usage{
writeaccessdatabase(data, database, sqtable, append = TRUE)
}
\arguments{
\item{data}{a data.frame to write to an Access database.}

\item{database}{file path to the Access database.}

\item{sqtable}{name of the table in the database.}

\item{append}{a logical indicating whether to append to the table or overwrite it. Default is TRUE.}
}
\description{
A function that writes data to an Access database using the RODBC package.
}
\keyword{access}
\keyword{database}
\keyword{export}
\keyword{read}
\keyword{save}
