% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratificationperiod.R
\name{stratificationperiod}
\alias{stratificationperiod}
\title{Lake Stratification Periods}
\usage{
stratificationperiod(tempdata, consecutivedays = 10)
}
\arguments{
\item{tempdata}{a table with "Date","Depth", and "Temperature" fields.}

\item{consecutivedays}{the number of consecutive days of uninterrupted stratification required to consider the lake stratified for the season. Default is 10 days.}
}
\value{
a data.frame indicating the start (S) and end (E) stratification dates.
}
\description{
This function calculates the start and end of lake thermal stratification using daily lake temperature profile data. In this calculation, a lake is considered stratified when there is a difference of one degree C between one meter depth intervals anywhere within the water column.
}
\examples{
\dontrun{
x <-stratificationperiod(
      tempdata = temperaturedata,
      consecutivedays = 15
      )
}
}
\keyword{lake}
\keyword{profile}
\keyword{stratification}
\keyword{temperature}
\keyword{thermal}
