% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterleveldownload.R
\name{waterleveldownload}
\alias{waterleveldownload}
\title{Download Water Level Data}
\usage{
waterleveldownload(lakeid)
}
\arguments{
\item{lakeid}{Minnesota lake identifier (DOWLKNUM) for the lake of interest.}
}
\value{
a data.frame with water level data
examples
x <- waterleveldownload("11041300")
}
\description{
This function downloads water level data from the MNDNR Lakefinder website and converts the elevation to meters. The data are already formatted into the
Sentinel Lakes format.
}
\keyword{Minnesota}
\keyword{data}
\keyword{levels}
\keyword{water}
