% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqmonthtrendstats.R
\name{wqmonthtrendstats}
\alias{wqmonthtrendstats}
\title{Calculate Water Quality Linear Trend Statistics}
\usage{
wqmonthtrendstats(statdata, logtransform = FALSE)
}
\arguments{
\item{statdata}{a water quality data.frame processed through the wqmonthtable() function.}

\item{logtransform}{a logical indicating whether the statistics should be calculated using the natural log of the parameter values. Defaults to FALSE.}
}
\value{
a data.frame with statistics for each parameter in the water quality data.frame input.
}
\description{
This function calculates linear trend statistics using data download from the MNPCA using the output of the wqmonthtable() function.
}
\seealso{
Other Water Quality: \code{\link{wqdatadownload}},
  \code{\link{wqmonthtable}},
  \code{\link{wqmonthtrendplots}},
  \code{\link{wqparameters}}, \code{\link{wqstations}}
}
\concept{Water Quality}
\keyword{analysis}
\keyword{minnesota}
\keyword{mnpca}
\keyword{pollution}
\keyword{quality}
\keyword{statistics}
\keyword{trend}
\keyword{water}
